Add back window.performance.memory attribute (for YouTube TV)
This provides duplicate functionality with upstream's
window.internals.memoryInfo(), but it is needed to be compliant with
YouTube TV requirements.

This was initially added by Neeraj Ravindra Deshpande, but was
reverted because of conflicts with upstream changes.

The new commit implements the attribute by copying
testing/MemoryInfo.* as page/PerformanceMemoryInfo.*, with a different
name to avoid build conflicts.