[GStreamer] add basic constraints in extendedSupportsType() for isTypeSupported
This compares width/heigth/framerate/bitrate/channels to maximum values.
Negative values are marked as unsupported.
The current code does not handle invalid parameters that are not integers,
which will be accepted.

Current max values are:
- width: 3840 for VP9, 1920 otherwise
- height: 2160 for VP9, 1080 otherwise
- framerate: 60 fps
- bitrate: 50 Mbps (arbitraty)
- channels: 2 (AAC codec limitation)

cryptoblockformat is currently not supported, we should add support
for "subsample" when VP9/Widevine support gets ready.