[MSE][GSTREAMER] Calculation rounding error in the DTS (DecodeTimestamp)
When the video display resolution changes, there is a calculation rounding error in
the DTS of the first buffer after the resolution is switched.
Thus, the MSE algorithm deletes this buffer when it checks if the current DTS is
greater or equal to the last DTS plus its duration.

The fix consists to use also the PTS value in MSE algorithm when it does this check.